Created a single lib.rs file
