--- programs/anchor-escrow/src/contexts/mod.rs
@@ -1,8 +1,0 @@
-pub mod make;
-pub use make::*;
-
-pub mod refund;
-pub use refund::*;
-
-pub mod take;
-pub use take::*;
