--- programs/anchor-escrow/src/state.rs
@@ -1,12 +1,0 @@
-use anchor_lang::prelude::*;
-
-#[account]
-#[derive(InitSpace)]
-pub struct Escrow {
-    pub seed: u64,
-    pub maker: Pubkey,
-    pub mint_a: Pubkey,
-    pub mint_b: Pubkey,
-    pub receive: u64,
-    pub bump: u8,
-}
